LG-9825: refactor to extends rails7 ComparisonValidator.
